Removed: slide-enter animation now static
Previously footer selects all items on the page and tries to enter as latest element. But this solution is not worked since selection gathered before main slot changing in layout. That causes incorrect stage for footer (too slow for short pages after long)

Now is using static 200ms which must be great for general purpose